Adding lib_discretization, lib_algebra and a test application named 'disc_test_app'. Both libraries are in an experimental status and all concepts and interfaces may be changed in the future without any announcement. Only internal testing is intended, external usage is strongly discouraged. Therefore, the default build process (using cmake) does exclude both libraries. To include the experimental libraries in the build process use 'cmake -DWITH_ALGEBRA'.

+++ ugbase/common/types.h
@@ -46,7 +46,7 @@ namespace ugtypes
 typedef unsigned int uint;
 typedef ugtypes::uint8_t byte;
 typedef ugtypes::uint32_t uint32;
-typedef float number;
+typedef double number;
 
 
 #endif
